+++ aiopyramid/websocket/config/__init__.py
@@ -1,0 +1,13 @@
+__all__ = []
+
+try:
+    from .uwsgi import *
+    __all__.append('UWSGIWebsocketMapper')
+except ImportError:
+    pass
+
+try:
+    from .gunicorn import *
+    __all__.append('WebsocketMapper')
+except ImportError:
+    pass