+++ aiopyramid/websocket/exceptions.py
@@ -1,0 +1,6 @@
+
+import greenlet
+
+
+class WebsocketClosed(greenlet.GreenletExit):
+    pass