+++ aiopyramid/websocket/view.py
@@ -1,23 +1,19 @@
-
 import asyncio
-
-from .config import UWSGIWebsocketMapper
 
 
 class WebsocketConnectionView:
     """ :term:`view_callable` for websocket connections. """
-
-    __view_mapper__ = UWSGIWebsocketMapper
 
     def __init__(self, context, request):
         self.context = context
         self.request = request
 
     @asyncio.coroutine
-    def __call__(self):
+    def __call__(self, ws):
+        self.ws = ws
         yield from self.on_open()
         while True:
-            message = yield from self._sock.recv()
+            message = yield from self.ws.recv()
             if message is None:
                 yield from self.on_close()
                 break
@@ -25,20 +21,27 @@ class WebsocketConnectionView:
 
     @asyncio.coroutine
     def send(self, message):
-        print('sending ', message)
-        yield from self._sock.send(message)
+        yield from self.ws.send(message)
 
     @asyncio.coroutine
     def on_message(self, message):
-        print('got ', message)
-        yield from self.send(message)
-
-    @asyncio.coroutine
-    def on_open(self):
-        print('opening')
+        """
+        Callback called when a message is received.
+        Default is a noop.
+        """
         pass
 
     @asyncio.coroutine
+    def on_open(self):
+        """
+        Callback called when the connection is first established.
+        Default is a noop.
+        """
+
+    @asyncio.coroutine
     def on_close(self):
-        print('closing')
+        """
+        Callback called when the connection is closed.
+        Default is a noop.
+        """
         pass